создана структура компонента с информацией о маршруте
