+++ src/mock/trip-point.js
@@ -1,6 +1,5 @@
-import dayjs from 'dayjs';
-import {DESTINATIONS, TYPES, MIN_PRICE, MAX_PRICE, DESTINATION_DESCRIPTIONS, RANDOM_MIN_DAY, RANDOM_MAX_DAY, RANDOM_MIN_TIME, RANDOM_MAX_TIME, MIN_OFFER_COUNT, MAX_OFFER_COUNT, MIN_COUNT_PHOTOS, MAX_COUNT_PHOTOS, MAX_NUMBER_PHOTO} from '../const.js';
-import {getRandomInteger} from '../utils.js';
+import {TYPES, MIN_PRICE, MAX_PRICE, DESTINATION_DESCRIPTIONS, MIN_OFFER_COUNT, MAX_OFFER_COUNT, MIN_COUNT_PHOTOS, MAX_COUNT_PHOTOS, MAX_NUMBER_PHOTO} from '../const.js';
+import {getRandomInteger, getRandomDayAndTime, getRandomDestination} from '../utils.js';
 import {generateOffer} from './offer.js';
 
 /**
@@ -13,15 +12,6 @@ export const getRandomPointType = () => {
 };
 
 /**
- * Функция генерации случайного напрвления для точки маршрута
- * @returns - строка с названием направления
- */
-const getRandomDestination = () => {
-  const randomIndex = getRandomInteger(0, DESTINATIONS.length - 1);
-  return DESTINATIONS[randomIndex];
-};
-
-/**
  * Функция генерации случайного описания о месте назначения точки маршрута
  * @returns  - строка с описанием места назначения
  */
@@ -30,15 +20,7 @@ const getRandomDestinationDescription = () => {
   return DESTINATION_DESCRIPTIONS[randomIndex];
 };
 
-/**
- * Функция генерация случайной даты со временем
- * @returns - случайная дата и время
- */
-const getRandomDayAndTime = () => {
-  const randomDate = getRandomInteger(RANDOM_MIN_DAY, RANDOM_MAX_DAY);
-  const randomTime = getRandomInteger(RANDOM_MIN_TIME, RANDOM_MAX_TIME);
-  return dayjs().add(randomDate, 'day').add(randomTime, 'minute').toDate();
-};
+
 /**
  * Функция генерация адреса случайной фотографии
  * @returns - строка с адресом случайного фото
